# schema/openapi/openapi.yaml
openapi: '3.0.3'
info:
  title: Backbone API
  version: '1.0.0'
  description: |
    Auto-generated by backbone-cli.
    This API provides 11 standard CRUD endpoints for each model.
servers:
  - url: http://localhost:3000
    description: Development server
...
paths:
...
  /api/v1/party_contacts:
    get:
      tags:
        - PartyContacts
      summary: List PartyContacts
      description: Get a paginated list of PartyContacts with optional filtering and sorting
      operationId: listPartyContacts
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/PerPageParam'
        - $ref: '#/components/parameters/SortByParam'
        - $ref: '#/components/parameters/SortOrderParam'
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PartyContactList'
        '401':
          $ref: '#/components/responses/Unauthorized'

    post:
      tags:
        - PartyContacts
      summary: Create PartyContact
      description: Create a new PartyContact
      operationId: createPartyContact
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreatePartyContactRequest'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PartyContact'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
  /api/v1/party_emails:
    get:
      tags:
        - PartyEmails
      summary: List PartyEmails
      description: Get a paginated list of PartyEmails with optional filtering and sorting
      operationId: listPartyEmails
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/PerPageParam'
        - $ref: '#/components/parameters/SortByParam'
        - $ref: '#/components/parameters/SortOrderParam'
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PartyEmailList'
        '401':
          $ref: '#/components/responses/Unauthorized'

    post:
      tags:
        - PartyEmails
      summary: Create PartyEmail
      description: Create a new PartyEmail
      operationId: createPartyEmail
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreatePartyEmailRequest'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PartyEmail'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
components:
  schemas:
...
    PartyContact:
      type: object
      required:
        - id
        - party_id
        - company_id
        - name
        - is_primary
        - metadata
      properties:
        id:
          type: string
          format: uuid
        party_id:
          type: string
          format: uuid
        company_id:
          type: string
          format: uuid
        name:
          type: string
          maxLength: 150
        job_title:
          nullable: true
          type: string
          maxLength: 120
        department:
          nullable: true
          type: string
          maxLength: 120
        email:
          nullable: true
          type: string
          maxLength: 150
        phone:
          nullable: true
          type: string
          maxLength: 30
        is_primary:
          type: boolean
        metadata:
          type: object
...
    CreatePartyContactRequest:
      type: object
      required:
        - party_id
        - company_id
        - name
        - is_primary
        - metadata
      properties:
        party_id:
          type: string
          format: uuid
        company_id:
          type: string
          format: uuid
        name:
          type: string
          maxLength: 150
        job_title:
          nullable: true
          type: string
          maxLength: 120
        department:
          nullable: true
          type: string
          maxLength: 120
        email:
          nullable: true
          type: string
          maxLength: 150
        phone:
          nullable: true
          type: string
          maxLength: 30
        is_primary:
          type: boolean
        metadata:
          type: object
...
    PartyContactList:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/PartyContact'
        meta:
          $ref: '#/components/schemas/PaginationMeta'
    PartyEmail:
      type: object
      required:
        - id
        - party_id
        - company_id
        - label
        - email
        - is_primary
        - metadata
      properties:
        id:
          type: string
          format: uuid
        party_id:
          type: string
          format: uuid
        company_id:
          type: string
          format: uuid
        label:
          type: string
          maxLength: 30
        email:
          type: string
          maxLength: 150
        is_primary:
          type: boolean
        metadata:
          type: object
...
    CreatePartyEmailRequest:
      type: object
      required:
        - party_id
        - company_id
        - label
        - email
        - is_primary
        - metadata
      properties:
        party_id:
          type: string
          format: uuid
        company_id:
          type: string
          format: uuid
        label:
          type: string
          maxLength: 30
        email:
          type: string
          maxLength: 150
        is_primary:
          type: boolean
        metadata:
          type: object
...
    PartyEmailList:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/PartyEmail'
        meta:
          $ref: '#/components/schemas/PaginationMeta'
...
    PaginationMeta:
      type: object
      properties:
        total:
          type: integer
          example: 100
        page:
          type: integer
          example: 1
        per_page:
          type: integer
          example: 20
        total_pages:
          type: integer
          example: 5
    Error:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: string
          example: VALIDATION_ERROR
        message:
          type: string
          example: Invalid input data
        details:
          type: object
          additionalProperties: true
  parameters:
...
    PageParam:
      name: page
      in: query
      schema:
        type: integer
        minimum: 1
        default: 1
    PerPageParam:
      name: per_page
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
    SortByParam:
      name: sort_by
      in: query
      schema:
        type: string
    SortOrderParam:
      name: sort_order
      in: query
      schema:
        type: string
        enum: [asc, desc]
        default: asc
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
security:
  - bearerAuth: []