# src/presentation/http/party_contact_handler.rs
//! PartyContact REST handlers
//!
//! Generated by metaphor-schema. Do not edit manually.
//!
//! Uses Axum and backbone-core's BackboneCrudHandler for all 12 CRUD endpoints.

use std::collections::HashMap;
use std::sync::Arc;

use axum::Router;
use serde::{Deserialize, Serialize};
use uuid::Uuid;

// Backbone framework imports
use backbone_core::http::BackboneCrudHandler;

// Auth integration (optional)
#[cfg(feature = "auth")]
use backbone_auth::middleware::AuthContext;
#[cfg(feature = "auth")]
use backbone_auth::AuthMiddleware;

// Domain imports
use crate::domain::entity::*;
use crate::application::service::{PartyContactService, ServiceError};

// DTO imports
use crate::presentation::dto::{CreatePartyContactDto, UpdatePartyContactDto, PatchPartyContactDto, PartyContactResponseDto};


/// Application error type
#[derive(Debug, thiserror::Error)]
pub enum PartyContactError {
    #[error("Not found: {0}")]
    NotFound(String),
    #[error("Validation error: {0}")]
    Validation(String),
    #[error("Database error: {0}")]
    Database(String),
    #[error("Internal error: {0}")]
    Internal(String),
}

impl From<ServiceError> for PartyContactError {
    fn from(err: ServiceError) -> Self {
        match err {
            ServiceError::NotFound => Self::NotFound(err.to_string()),
            ServiceError::Validation(ref msg) => Self::Validation(msg.clone()),
            ServiceError::AlreadyExists(ref msg) => Self::Validation(msg.clone()),
            ServiceError::Repository(ref e) => Self::Database(e.to_string()),
            ServiceError::Internal(ref msg) => Self::Internal(msg.clone()),
        }
    }
}

impl axum::response::IntoResponse for PartyContactError {
    fn into_response(self) -> axum::response::Response {
        use axum::http::StatusCode;
        use axum::Json;

        let (status, code) = match &self {
            Self::NotFound(_) => (StatusCode::NOT_FOUND, "PARTYCONTACT_NOT_FOUND"),
            Self::Validation(_) => (StatusCode::BAD_REQUEST, "PARTYCONTACT_VALIDATION_ERROR"),
            Self::Database(_) => (StatusCode::INTERNAL_SERVER_ERROR, "PARTYCONTACT_DATABASE_ERROR"),
            Self::Internal(_) => (StatusCode::INTERNAL_SERVER_ERROR, "PARTYCONTACT_INTERNAL_ERROR"),
        };

        let body = serde_json::json!({
            "success": false,
            "error": code,
            "message": self.to_string(),
        });

        (status, Json(body)).into_response()
    }
}

// =============================================================================
// Route Configuration
// =============================================================================

/// Create Axum router with all 16 Backbone endpoints for PartyContact.
///
/// # Routes
///
/// | Method | Path | Description |
/// |--------|------|-------------|
/// | GET | /party_contacts | List with pagination |
/// | POST | /party_contacts | Create new |
/// | GET | /party_contacts/:id | Get by ID |
/// | PUT | /party_contacts/:id | Full update |
/// | PATCH | /party_contacts/:id | Partial update |
/// | DELETE | /party_contacts/:id | Soft delete |
/// | POST | /party_contacts/bulk | Bulk create |
/// | POST | /party_contacts/upsert | Upsert |
/// | GET | /party_contacts/trash | List deleted |
/// | POST | /party_contacts/:id/restore | Restore |
/// | DELETE | /party_contacts/empty | Empty trash |
/// | GET | /party_contacts/:id/deleted | Get deleted by ID |
/// | DELETE | /party_contacts/trash/:id | Permanent delete from trash |
/// | GET | /party_contacts/count | Count active entities |
/// | GET | /party_contacts/trash/count | Count deleted entities |
///
/// # Example
///
/// ```text
/// let service = Arc::new(PartyContactService::with_repository(repository));
/// let router = create_party_contact_routes(service);
/// ```
pub fn create_party_contact_routes(service: Arc<PartyContactService>) -> Router {
    BackboneCrudHandler::<PartyContactService, PartyContact, CreatePartyContactDto, UpdatePartyContactDto, PartyContactResponseDto>::routes(
        service,
        "/party_contacts",
    )
}

/// Create Axum router with only the read (GET) endpoints for PartyContact.
///
/// Safe for public, unauthenticated exposure (e.g., reference data).
/// Mutations must be served separately via `create_party_contact_write_routes`,
/// typically wrapped in an auth middleware layer.
pub fn create_party_contact_read_routes(service: Arc<PartyContactService>) -> Router {
    BackboneCrudHandler::<PartyContactService, PartyContact, CreatePartyContactDto, UpdatePartyContactDto, PartyContactResponseDto>::read_routes(
        service,
        "/party_contacts",
    )
}

/// Create Axum router with only the write (mutation) endpoints for PartyContact.
///
/// These routes must NOT be publicly exposed. Wrap them with an auth
/// middleware before nesting into the application router.
///
/// # This is unguarded generic CRUD, not a validated write path
///
/// These are plain create/update/patch/delete mutations over the entity row —
/// they bypass all business invariants. If the module exposes a validated write
/// service (e.g. a command router over its domain engine), serve THAT instead
/// for any mutation that must respect domain rules.
pub fn create_party_contact_write_routes(service: Arc<PartyContactService>) -> Router {
    BackboneCrudHandler::<PartyContactService, PartyContact, CreatePartyContactDto, UpdatePartyContactDto, PartyContactResponseDto>::write_routes(
        service,
        "/party_contacts",
    )
}

/// Create authenticated routes with auth middleware.
///
/// Requires the `auth` feature flag. The `AuthMiddleware` implementation
/// is responsible for extracting and validating tokens, then providing
/// an `AuthContext` via request extensions.
#[cfg(feature = "auth")]
pub fn create_protected_party_contact_routes<A: AuthMiddleware + Send + Sync + 'static>(
    service: Arc<PartyContactService>,
    auth: Arc<A>,
) -> Router {
    use axum::middleware;
    use axum::response::IntoResponse;

    let auth_layer = auth.clone();
    create_party_contact_routes(service)
        .layer(middleware::from_fn(move |mut req: axum::extract::Request, next: axum::middleware::Next| {
            let auth = auth_layer.clone();
            async move {
                let token = req.headers()
                    .get(axum::http::header::AUTHORIZATION)
                    .and_then(|h| h.to_str().ok())
                    .and_then(|raw| raw.strip_prefix("Bearer ").or_else(|| raw.strip_prefix("bearer ")))
                    .unwrap_or("");
                match auth.authenticate(token).await {
                    Ok(ctx) => {
                        req.extensions_mut().insert(ctx);
                        next.run(req).await
                    }
                    Err(_) => {
                        (axum::http::StatusCode::UNAUTHORIZED,
                         axum::Json(serde_json::json!({
                             "success": false,
                             "error": "unauthorized",
                             "message": "Authentication required"
                         }))
                        ).into_response()
                    }
                }
            }
        }))
}